Allows asNavFor to apply to multiple slicks
Changes the asNavFor function to allow one slick instance to control multiple instances. It excludes itself from the list to avoid an endless loop situation. 

The function doesn't test for null explicitly anymore, as passing null to jQuery will result in a no-op anyway. If this.options.asNavFor is a valid jQuery selector, it will run slideHandler() on each instance of slick that it finds on each matching node that jQuery finds.